refactor[api]: external api calls
